+++ AngularClient/projects/forcrowd-backbone/src/lib/auth.service.ts
@@ -4,7 +4,7 @@ import { EntityQuery, EntityState, MergeStrategy } from "breeze-client";
 import { throwError as observableThrowError, of as observableOf, Observable, Subject } from "rxjs";
 import { catchError, mergeMap, map } from "rxjs/operators";
 
-import { AppSettings } from "../../app-settings/app-settings";
+import { AppSettings } from "./app-settings/app-settings";
 import { Role } from "./entities/role";
 import { User } from "./entities/user";
 import { UserRole } from "./entities/user-role";
@@ -12,7 +12,7 @@ import { AppHttpClient } from "./app-http-client/app-http-client.module";
 import { AppEntityManager } from "./app-entity-manager.service";
 import { NotificationService } from "./notification.service";
 import { Token } from "./token";
-import { getUniqueUserName } from "../shared/utils";
+import { getUniqueUserName } from "./shared/utils";
 
 @Injectable()
 export class AuthService {