+++ AngularClient/src/dev-all/odata-element-field.component.ts
@@ -4,11 +4,7 @@ import { Observable } from "rxjs";
 import { mergeMap, map } from "rxjs/operators";
 
 import { AppSettings } from "../app-settings/app-settings";
-import { Element } from "../main/core/entities/element";
-import { Project } from "../main/core/entities/project";
-import { User } from "../main/core/entities/user";
-import { AuthService } from "../main/core/core.module";
-import { getUniqueValue } from "../main/shared/utils";
+import { Element, Project, User, AuthService, getUniqueValue } from "forcrowd-backbone";
 
 @Component({
   selector: "odata-element-field",