+++ AngularClient/src/main/account/account-overview.component.ts
@@ -1,11 +1,12 @@
 import { Component } from "@angular/core";
 
-import { User } from "../core/entities/user";
+import { User } from "forcrowd-backbone";
 import { AuthService } from "../core/core.module";
 
 @Component({
   selector: "account-overview",
-  templateUrl: "account-overview.component.html"
+  templateUrl: "account-overview.component.html",
+  styleUrls: ["account-overview.component.css"]
 })
 export class AccountOverviewComponent {