+++ AngularClient/src/main/project/project-manager.component.ts
@@ -1,8 +1,7 @@
 import { Component, OnInit } from "@angular/core";
 import { ActivatedRoute, Router } from "@angular/router";
 
-import { Project } from "../core/entities/project";
-import { User } from "../core/entities/user";
+import { Project, User } from "forcrowd-backbone";
 import { ProjectService } from "../core/core.module";
 
 @Component({