+++ AngularClient/src/main/project/remove-confirm.component.ts
@@ -1,0 +1,8 @@
+import { Component } from "@angular/core";
+
+@Component({
+  selector: "profile-remove-project",
+  templateUrl: "remove-confirm.component.html",
+  styleUrls: ["remove-confirm.component.css"]
+})
+export class RemoveConfirmComponent { }