+++ AngularClient/src/main/user/user.module.ts
@@ -1,7 +1,8 @@
 import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
+import { FlexLayoutModule } from "@angular/flex-layout";
 
-import { SharedModule } from "../shared/shared.module";
+import { SharedModule } from "forcrowd-backbone";
 
 import { ProfileComponent } from "./profile.component";
 import { ProfileRemoveProjectComponent } from "./profile-remove-project.component";
@@ -25,6 +26,7 @@ const userRoutes: Routes = [
   ],
   imports: [
     SharedModule,
+    FlexLayoutModule,
     RouterModule.forChild(userRoutes),
   ],
   providers: [