+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ### Changelog
 
+**0.9.0**
+
+ * Angular library has been added "forcrowd-backbone" for many service
+ * AngularClient - Backbone new UI has been added
+  * Minor update
+
 **0.8.1**
 
 * AngularClient - 'Todo App' template fix